Fix float32 overflow in relative_logit_error on microsoft/DialoGPT-small
Discovered while running Experiment 29 (the fourth pretrained model,
chosen to isolate scale from training procedure): DialoGPT-small's raw
logits are large enough that the naive sum-of-squares norm in
run_layer_experiment silently overflowed float32, producing NaN for
relative_logit_error on every single row of the initial smoke-test run
(confirmed via numpy's own "overflow encountered in dot" RuntimeWarning).

Fix: cast logits to float64 before computing the norm. This is a
correctness fix in shared code (atlas_nn/stage_b/experiment.py,
used by every experiment since Stage B), not model-specific -- float64
gives numerically identical results to float32 on every model whose
logits don't approach overflow range, so no prior experiment's results
(distilgpt2, gpt2, gpt2-medium, Stage B, Stage C-lite) are affected or
need rerunning. Verified directly: the same row that previously returned
NaN now returns a sane 0.0036.

Full suite still green: 71 passed, 1 warning (pre-existing, unrelated).
The DialoGPT-small smoke test itself needs to be rerun from scratch,
since its first attempt's results are all NaN and unusable.

+++ atlas_nn/stage_b/experiment.py
@@ -77,8 +77,14 @@ def run_layer_experiment(
     finally:
         load_snapshot(model, original_state)
 
-    logits_orig = baseline_eval["logits"]
-    logits_sub = substituted_eval["logits"]
+    # float64, not the logits' native float32: some real pretrained models
+    # (e.g. microsoft/DialoGPT-small) produce raw logits large enough that
+    # a naive sum-of-squares norm overflows float32 (silently returning NaN)
+    # well within its representable range in float64. This changes no
+    # result on any model whose logits don't approach that scale -- it only
+    # avoids a precision artifact, not a change in what's measured.
+    logits_orig = baseline_eval["logits"].astype(np.float64)
+    logits_sub = substituted_eval["logits"].astype(np.float64)
     logit_diff_norm = float(np.linalg.norm(logits_sub - logits_orig))
     logit_orig_norm = float(np.linalg.norm(logits_orig))
     relative_logit_error = logit_diff_norm / logit_orig_norm if logit_orig_norm > 0 else logit_diff_norm
